3498 - update Data source validations
