+++ src/client/pages/Section/Descriptions/NationalDataDescriptions/DataSources/Columns/Variables.tsx
@@ -1,8 +1,12 @@
-import React from 'react'
+import React, { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
+
+import classNames from 'classnames'
+import { Objects } from 'utils/objects'
 
 import { DataSource, Labels, SectionName } from 'meta/assessment'
 import { DataSourceDescription } from 'meta/assessment/description/nationalDataDataSourceDescription'
+import { TooltipId } from 'meta/tooltip'
 
 import { DataCell } from 'client/components/DataGrid'
 import Select from 'client/components/Inputs/Select'
@@ -60,10 +64,22 @@ const VariablesSelect: React.FC<Props> = (props) => {
 const Variables: React.FC<Props & { lastRow: boolean }> = (props) => {
   const { dataSource, disabled, meta, lastRow, sectionName } = props
 
+  const { t } = useTranslation()
+
+  const validationError = useMemo(() => {
+    return !dataSource.placeholder && Objects.isEmpty(dataSource.variables)
+  }, [dataSource.placeholder, dataSource.variables])
+
   const Component = meta.table?.variables?.length > 0 ? VariablesSelect : VariablesText
 
   return (
-    <DataCell editable={!disabled} lastRow={lastRow}>
+    <DataCell
+      className={classNames({ 'validation-error': validationError })}
+      data-tooltip-content={validationError ? t('generalValidation.notEmpty') : ''}
+      data-tooltip-id={TooltipId.error}
+      editable={!disabled}
+      lastRow={lastRow}
+    >
       <Component dataSource={dataSource} disabled={disabled} meta={meta} sectionName={sectionName} />
     </DataCell>
   )
