+++ src/client/pages/Section/Descriptions/NationalDataDescriptions/DataSources/Columns/YearForDataSource.tsx
@@ -1,6 +1,11 @@
-import React from 'react'
+import React, { useMemo } from 'react'
+import { useTranslation } from 'react-i18next'
+
+import classNames from 'classnames'
+import { Objects } from 'utils/objects'
 
 import { DataSource, SectionName } from 'meta/assessment'
+import { TooltipId } from 'meta/tooltip'
 
 import { DataCell } from 'client/components/DataGrid'
 import TextArea from 'client/components/Inputs/TextArea'
@@ -17,12 +22,23 @@ type Props = {
 const YearForDataSource: React.FC<Props> = (props: Props) => {
   const { dataSource, disabled, lastRow, sectionName } = props
 
+  const { t } = useTranslation()
   const onChange = useOnChange({ sectionName, dataSource })
+
+  const validationError = useMemo(() => {
+    return !dataSource.placeholder && Objects.isEmpty(dataSource.year)
+  }, [dataSource.placeholder, dataSource.year])
 
   const _onChange: React.ChangeEventHandler<HTMLTextAreaElement> = (event) => onChange('year', event.target.value)
 
   return (
-    <DataCell editable={!disabled} lastRow={lastRow}>
+    <DataCell
+      className={classNames({ 'validation-error': validationError })}
+      data-tooltip-content={validationError ? t('generalValidation.notEmpty') : ''}
+      data-tooltip-id={TooltipId.error}
+      editable={!disabled}
+      lastRow={lastRow}
+    >
       <TextArea disabled={disabled} onChange={_onChange} value={dataSource.year} />
     </DataCell>
   )
